remove isdigit

+++ net.c
@@ -256,7 +256,7 @@ validate_ipv4(const char *ip)
     while ((p = seperate_string(ip, ".", &len, cnt++)) != NULL) {
         sum = 0;
         for (i = 0; i < len; ++i) {
-            if (!isdigit(p[i]))
+            if (!isdigit((int)p[i]))
                 return 0;
             sum = sum * 10 + p[i] - '0';
         }
